ssecuring the app

+++ taskmate/todolist_app/views.py
@@ -16,7 +16,9 @@ def todolist(request):
     if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
         messages.success(request, ("New Task Added!"))
         return redirect('todolist')
     else:
@@ -30,7 +32,10 @@ def todolist(request):
 @login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)  # pk is for primary key
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request, ("Access Restricted, You Are Not Allowed."))
     return redirect('todolist')
 
 
@@ -51,8 +56,11 @@ def edit_task(request, task_id):
 @login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)  # pk is for primary key
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request, ("Access Restricted, You Are Not Allowed."))
     return redirect('todolist')
 
 
@@ -71,6 +79,7 @@ def index(request):
     return render(request, 'index.html', context)  # {'index_text': context}
 
 
+@login_required
 def contact(request):
     context = {
         'contact_text': 'Welcome to Contact page',
